# api-gateway/internal/http/docs/swagger.yaml
definitions:
...
  genproto.UserChangePasswordReq:
    properties:
      email:
        type: string
      new_password:
        type: string
    type: object
...
  genproto.UserGetRes:
    properties:
      dob:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: string
      last_name:
        type: string
      pfp_url:
        type: string
    type: object
...
  genproto.UserUpdateBody:
    properties:
      dob:
        type: string
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /delete-account:
    delete:
      consumes:
      - application/json
      description: Deletes the user's account permanently.
      produces:
      - application/json
      responses:
        "200":
          description: Account deleted successfully
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete user account
      tags:
      - 03-User-profile
...
  /profile:
    get:
      consumes:
      - application/json
      description: Get the profile of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.UserGetRes'
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - 03-User-profile
...
  /set-pfp:
    post:
      consumes:
      - multipart/mixed
      description: Adds a profile image to user.
      parameters:
      - description: Profile image
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Profile image is added
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Set a profile picture
      tags:
      - 03-User-profile
  /user:
    put:
      consumes:
      - application/json
      description: Update user's data
      parameters:
      - description: User Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/genproto.UserUpdateBody'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Error updating User
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update user's profile. In order not to update any field, leave it as
        it is like "string" or ""
      tags:
      - 03-User-profile
  /user-password:
    put:
      consumes:
      - application/json
      description: Updates the password to new one
      parameters:
      - description: Change Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/genproto.UserChangePasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: Password successfully updated
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "401":
          description: Incorrect verification code
          schema:
            type: string
        "404":
          description: Verification code expired or email not found
          schema:
            type: string
        "500":
          description: Error updating password
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - 03-User-profile
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...